Refactor PredisConnection and transaction() implementation
This rewrites the implementation of the transaction() method to use the
Predis MULTI/EXEC abstraction and adds facilities for high-availability
using a routine similar to the implementation in Predis.

Additionally, we rename the PredisSentinelConnection class to PredisConnection
for consistency with the existing PredisConnector and Laravel's Connector and
Connection classes. The explicit use of Sentinel in the name is arguably
welcome for disambiguation between the framework and package classes, but in
this case, these classes are intended for internal consumption so we'll remove
Sentinel for more concise names and to reflect that these are analogous to
their framework counterparts.

Signed-off-by: Cy Rossignol <cy@rossignols.me>